+++ src/main/java/org/opensearch/knn/index/codec/nativeindex/NativeIndexBuildStrategyFactory.java
@@ -7,14 +7,18 @@ package org.opensearch.knn.index.codec.nativeindex;
 
 import org.apache.lucene.index.FieldInfo;
 import org.opensearch.index.IndexSettings;
+import org.opensearch.knn.common.featureflags.KNNFeatureFlags;
 import org.opensearch.knn.index.codec.nativeindex.remote.RemoteIndexBuildStrategy;
 import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.index.vectorvalues.KNNVectorValues;
 import org.opensearch.repositories.RepositoriesService;
 
+import java.io.IOException;
 import java.util.function.Supplier;
 
 import static org.opensearch.knn.common.FieldInfoExtractor.extractKNNEngine;
 import static org.opensearch.knn.common.KNNConstants.MODEL_ID;
+import static org.opensearch.knn.index.codec.util.KNNCodecUtil.initializeVectorValues;
 
 /**
  * Creates the {@link NativeIndexBuildStrategy}
@@ -34,11 +38,18 @@ public final class NativeIndexBuildStrategyFactory {
     }
 
     /**
-     * Creates or returns the desired {@link NativeIndexBuildStrategy} implementation. Intended to be used by {@link NativeIndexWriter}
-     * @param fieldInfo
-     * @return
+     * @param fieldInfo             Field related attributes/info
+     * @param totalLiveDocs         Number of documents with the vector field. This values comes from {@link org.opensearch.knn.index.codec.KNN990Codec.NativeEngines990KnnVectorsWriter#flush}
+     *                              and {@link org.opensearch.knn.index.codec.KNN990Codec.NativeEngines990KnnVectorsWriter#mergeOneField}
+     * @param knnVectorValues       An instance of {@link KNNVectorValues} which is used to evaluate the size threshold KNN_REMOTE_VECTOR_BUILD_THRESHOLD
+     * @return                      The {@link NativeIndexBuildStrategy} to be used. Intended to be used by {@link NativeIndexWriter}
+     * @throws IOException
      */
-    public NativeIndexBuildStrategy getBuildStrategy(final FieldInfo fieldInfo) {
+    public NativeIndexBuildStrategy getBuildStrategy(
+        final FieldInfo fieldInfo,
+        final int totalLiveDocs,
+        final KNNVectorValues<?> knnVectorValues
+    ) throws IOException {
         final KNNEngine knnEngine = extractKNNEngine(fieldInfo);
         boolean isTemplate = fieldInfo.attributes().containsKey(MODEL_ID);
         boolean iterative = !isTemplate && KNNEngine.FAISS == knnEngine;
@@ -47,11 +58,15 @@ public final class NativeIndexBuildStrategyFactory {
             ? MemOptimizedNativeIndexBuildStrategy.getInstance()
             : DefaultIndexBuildStrategy.getInstance();
 
-        if (repositoriesServiceSupplier != null
+        initializeVectorValues(knnVectorValues);
+        long vectorBlobLength = ((long) knnVectorValues.bytesPerVector()) * totalLiveDocs;
+
+        if (KNNFeatureFlags.isKNNRemoteVectorBuildEnabled()
+            && repositoriesServiceSupplier != null
             && indexSettings != null
             && knnEngine.supportsRemoteIndexBuild()
-            && RemoteIndexBuildStrategy.shouldBuildIndexRemotely(indexSettings)) {
-            return new RemoteIndexBuildStrategy(repositoriesServiceSupplier, strategy);
+            && RemoteIndexBuildStrategy.shouldBuildIndexRemotely(indexSettings, vectorBlobLength)) {
+            return new RemoteIndexBuildStrategy(repositoriesServiceSupplier, strategy, indexSettings);
         } else {
             return strategy;
         }