+++ src/test/java/org/opensearch/knn/index/vectorvalues/TestVectorValues.java
@@ -22,6 +22,7 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 
+import static com.carrotsearch.randomizedtesting.RandomizedTest.randomByte;
 import static com.carrotsearch.randomizedtesting.RandomizedTest.randomFloat;
 
 public class TestVectorValues {
@@ -361,6 +362,14 @@ public class TestVectorValues {
         float[] data = new float[dimension];
         for (int i = 0; i < dimension; i++) {
             data[i] = randomFloat();
+        }
+        return data;
+    }
+
+    public static byte[] getRandomByteVector(int dimension) {
+        byte[] data = new byte[dimension];
+        for (int i = 0; i < dimension; i++) {
+            data[i] = randomByte();
         }
         return data;
     }